requirements.txt
pandas
numpy
re

yake
time
streamlit
Pillow

requests

streamlit
